refact: Refactored package structure to further implement game logic
