+++ app/src/main/AndroidManifest.xml
@@ -17,24 +17,24 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".GameBoardActivity"
+            android:name=".ui.GameBoardActivity"
             android:exported="false"/>
         <activity
-            android:name=".InLobbyActivity"
+            android:name=".ui.InLobbyActivity"
             android:exported="false" />
         <activity
-            android:name=".LobbyListsActivity"
+            android:name=".ui.LobbyListsActivity"
             android:exported="false"
             android:label="@string/title_activity_lobby_lists"
             android:theme="@style/Carcassonne" />
         <activity
-            android:name=".GameLobbyActivity"
+            android:name=".ui.GameLobbyActivity"
             android:exported="false" />
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.HomeActivity"
             android:exported="false" />
         <activity
-            android:name=".StartupActivity"
+            android:name=".ui.StartupActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
