+++ app/src/main/java/se2/carcassonne/ui/GameboardAdapter.java
@@ -1,10 +1,12 @@
-package se2.carcassonne;
+package se2.carcassonne.ui;
 
 import android.content.Context;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
+
+import se2.carcassonne.R;
 
 public class GameboardAdapter extends BaseAdapter {
