+++ app/src/main/res/layout/game_lobby_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/home_background"
-    tools:context=".GameLobbyActivity">
+    tools:context=".ui.GameLobbyActivity">
 
     <TextView
         android:id="@+id/textViewStartPlaying"
