+++ app/src/main/res/layout/home_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/startup_background"
-    tools:context=".HomeActivity">
+    tools:context=".ui.HomeActivity">
 
     <ImageView
         android:id="@+id/logo_animation"
