+++ app/src/main/res/layout/in_lobby_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/home_background"
-    tools:context=".InLobbyActivity">
+    tools:context=".ui.InLobbyActivity">
 
     <TextView
         android:id="@+id/textViewLobbyName"
