+++ app/src/main/res/layout/lobby_list_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/home_background"
-    tools:context=".HomeActivity">
+    tools:context=".ui.HomeActivity">
 
     <TextView
         android:id="@+id/textView"
