+++ app/src/main/res/layout/startup_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/startup_background"
-    tools:context=".StartupActivity">
+    tools:context=".ui.StartupActivity">
 
     <ImageView
         android:id="@+id/imageView"
